renaming CarouselSlider
