+++ src/components/ui/ImageGallery/ImageGalleryZoom.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const ImageGalleryZoom = () => {
+  return <span />
+}
+
+export default ImageGalleryZoom
