+++ src/components/ui/ImageGallery/index.tsx
@@ -1,3 +1,4 @@
-export { default as ImageGallery } from './ImageGallery'
+export { default as ImageGalleryProvider } from './ImageGalleryProvider'
 export { default as ImageGallerySelector } from './ImageGallerySelector'
-export { default as CarouselSlider } from './CarouselSlider'
+export { default as ImageGallery } from './ImageGallery'
+export { default as ImageGalleryZoom } from './ImageGalleryZoom'
